requirements.txt
streamlit
langchain
langchain_groq
langchain_community
gitpython
python-dotenv
faiss-cpu
openai
